move classes to dedicated folder
